tests renamed and reworked
