+++ src/category_class.py
@@ -11,10 +11,10 @@ class Category:
     def __init__(self, name, description, products):
         self.name = name
         self.description = description
-        self.__products = products
+        self.__products = products if products else []
 
         Category.category_count += 1
-        Category.product_count += len(products)
+        Category.product_count += len(self.__products)
 
     def add_product(self, new_product: Product):
         self.__products.append(new_product)
@@ -23,7 +23,7 @@ class Category:
     @property
     def products(self):
         product_str = ""
-        for product in self.products:
+        for product in self.__products:
             product_str += (
                 f"{product.name}, {product.price} руб. Остаток: {product.quantity} шт.\n"
             )
